handle slide cuepoints in simulive
