+++ src/providers/vod/index.ts
@@ -1,0 +1,2 @@
+export {HotspotLoader} from './hotspot-loader';
+export {ThumbLoader} from './thumb-loader';
